backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "faiss-cpu>=1.9.0.post1",
    "fastapi>=0.115.5",
    "langchain-ollama>=0.2.0",
    "langdetect>=1.0.9",
    "llama-index>=0.12.1",
    "llama-index-llms-openai>=0.3.1",
    "openai>=1.55.0",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.1",
    "llama-index-llms-gemini>=0.4.0",
    "llama-index-vector-stores-faiss>=0.3.0",
    "python-dotenv>=1.0.1",
]
